requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pika==1.3.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
websocket-client==1.8.0
wsproto==1.2.0
